# back-end/laravel/public/api-documentation/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Employment Api Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/employees": {
            "get": {
                "tags": [
                    "EmployeeController"
                ],
                "summary": "Obter dados de todos os colaboradores",
                "description": "end-point para obter os dados de todos os colaboradores",
                "operationId": "7fc06f144c1cd7925c5c391a6609f02d",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "pagina atual da paginacao",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna uma lista com objetos contendo os dados dos colaboradores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Fulano"
                                            },
                                            "surname": {
                                                "type": "string",
                                                "example": "De Tal"
                                            },
                                            "salary": {
                                                "type": "string",
                                                "example": 12334455690
                                            },
                                            "user_id": {
                                                "type": "number",
                                                "example": "5"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/employees/{id}": {
            "get": {
                "tags": [
                    "EmployeeController"
                ],
                "summary": "Obter dados de um colaborador especifico com base no id",
                "description": "end-point para obter os dados referentes a um determinado colaborador",
                "operationId": "d53e117f7e819b6f063fc96ea910eb75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do colaborador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna os dados do colaborador",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Fulano"
                                        },
                                        "surname": {
                                            "type": "string",
                                            "example": "De Tal"
                                        },
                                        "salary": {
                                            "type": "string",
                                            "example": 12334455690
                                        },
                                        "user_id": {
                                            "type": "number",
                                            "example": "5"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "EmployeeController"
                ],
                "summary": "Deletar conta do colaborador",
                "description": "end-point utilizado para deletar a propria conta colaborador (Apenas a conta proprietaria do token de autorizacao pode ser deletada).",
                "operationId": "4743ab328b7641983d97c047c641038f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do colaborador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbiden (Ao tentar apagar uma conta que nao seja a propria)"
                    },
                    "500": {
                        "description": "Interna Server Error"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Registrar um novo colaborador",
                "description": "end-point utilizado para registrar um novo colaborador",
                "operationId": "4d6e95be84db27edf40c5f8bf6cc295a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "cpf": {
                                        "type": "string",
                                        "example": "11222330188"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Fulano"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "De Tal"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "teste_usuario@example.org"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna dados do usuario e dados do colaborador"
                    },
                    "500": {
                        "description": "Interna Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "description": "Baerer token obtido na autenticação",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}